Create FinalSystemCode
This code is for the automated control valve created by the Coanda Corporation ENME 444 Fall 2020 Capstone group. The code is used to control a ball valve using a stepper motor as an actuator based on user input through buttons and data collected from a proximity sensor that is used to sense mass flow rate and is attached to the rotameter on the P.A. Hilton F840 Experimental Reaction Turbine. The code also displays the set and real mass airflow values to an LCD module attached to the turbine.